requirements.txt
asttokens==3.0.1
biopython==1.85
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
maturin==1.9.6
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
